Enhance Azure Pipelines for SQL database deployment (#15)
* Enhance Azure Pipelines for SQL database deployment

Updated `azure-pipelines-database.yml` to improve the build and deployment process. Added a new build stage for DACPAC creation and a deployment stage for DEV environment. Replaced manual SQLPackage execution with a built-in Azure DevOps task. Improved YAML structure and adjusted variables for better clarity and efficiency.

* Locate single DACPAC file before deployment

Updated `azure-pipelines-user-database.yml` to add a step that dynamically locates a single DACPAC file for deployment. This change replaces the hardcoded DACPAC path with a check for the existence of exactly one DACPAC file, improving error handling and robustness in the deployment process.